# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DopaFlow API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/review/cards": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "List Cards",
        "description": "List review cards due today.",
        "operationId": "list_cards_api_v2_review_cards_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewCardRead"
                  },
                  "type": "array",
                  "title": "Response List Cards Api V2 Review Cards Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Create Card",
        "description": "Create a review card.",
        "operationId": "create_card_api_v2_review_cards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCardCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCardRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/review/due": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Due Cards",
        "description": "List due cards for a specific deck.",
        "operationId": "due_cards_api_v2_review_due_get",
        "parameters": [
          {
            "name": "deck_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Deck Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewCardRead"
                  },
                  "title": "Response Due Cards Api V2 Review Due Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/review/decks": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Create Deck",
        "description": "Create a review deck.",
        "operationId": "create_deck_api_v2_review_decks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeckCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Deck Api V2 Review Decks Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/review/rate": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Rate Card",
        "description": "Rate a card and persist updated SM-2 scheduling metadata.",
        "operationId": "rate_card_api_v2_review_rate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRating"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCardRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeckCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type",
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "DeckCreate",
        "description": "Payload for creating a spaced repetition deck."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewCardCreate": {
        "properties": {
          "deck_id": {
            "type": "string",
            "title": "Deck Id"
          },
          "front": {
            "type": "string",
            "title": "Front"
          },
          "back": {
            "type": "string",
            "title": "Back"
          }
        },
        "type": "object",
        "required": [
          "deck_id",
          "front",
          "back"
        ],
        "title": "ReviewCardCreate",
        "description": "Payload for creating a review card."
      },
      "ReviewCardRead": {
        "properties": {
          "deck_id": {
            "type": "string",
            "title": "Deck Id"
          },
          "front": {
            "type": "string",
            "title": "Front"
          },
          "back": {
            "type": "string",
            "title": "Back"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "interval": {
            "type": "integer",
            "title": "Interval",
            "default": 0
          },
          "ease_factor": {
            "type": "number",
            "title": "Ease Factor",
            "default": 2.5
          },
          "next_review_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Review At"
          },
          "last_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Rating"
          },
          "lapse_count": {
            "type": "integer",
            "title": "Lapse Count",
            "default": 0
          },
          "reviews_done": {
            "type": "integer",
            "title": "Reviews Done",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "deck_id",
          "front",
          "back",
          "id"
        ],
        "title": "ReviewCardRead",
        "description": "Serialized review card including SM-2 scheduling fields."
      },
      "ReviewRating": {
        "properties": {
          "card_id": {
            "type": "string",
            "title": "Card Id"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "rating"
        ],
        "title": "ReviewRating",
        "description": "Payload for grading a review card."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}